implement 10.11

+++ core-rs/migrations/0013_properties.sql
@@ -2,7 +2,7 @@
 -- DAV custom properties set via PROPPATCH. Keyed by userid + path + name.
 
 CREATE TABLE IF NOT EXISTS oc_properties (
-    id            BIGINT       NOT NULL PRIMARY KEY,
+    id            INTEGER      NOT NULL PRIMARY KEY,
     userid        VARCHAR(64)  NOT NULL DEFAULT '',
     propertypath  VARCHAR(255) NOT NULL DEFAULT '',
     propertyname  VARCHAR(255) NOT NULL DEFAULT '',
